Fixed to double

+++ src/edu/uga/cs/rentaride/presentation/CreateVehicleType.java
@@ -111,7 +111,7 @@ public class CreateVehicleType
         priceS = req.getParameter( "membershipPrice" );
 
         try{
-            price = Integer.valueOf(priceS);
+            price = Double.valueOf(priceS);
         }catch(Exception e) {
 
         }
